Move barcode scanner code to task
